+++ packages/plugins/GeoLocation/src/store/actions.ts
@@ -8,11 +8,13 @@ import { Style, Icon } from 'ol/style'
 import * as Proj from 'ol/proj.js'
 import Geolocation from 'ol/Geolocation.js'
 import { transform as transformCoordinates } from 'ol/proj'
+import Overlay from 'ol/Overlay'
 import { GeoLocationState, GeoLocationGetters } from '../types'
 import geoLocationMarker from '../assets/geoLocationMarker'
+import { getTooltip } from '../utils/tooltip'
 
 const actions: PolarActionTree<GeoLocationState, GeoLocationGetters> = {
-  setupModule({ getters: { checkLocationInitially }, commit, dispatch }): void {
+  setupModule({ getters, commit, dispatch }): void {
     dispatch('addMarkerLayer')
 
     // NOTE: limited support across browsers
@@ -21,8 +23,34 @@ const actions: PolarActionTree<GeoLocationState, GeoLocationGetters> = {
         commit('setIsGeolocationDenied', true)
       }
     })
-    if (checkLocationInitially) {
+    if (getters.checkLocationInitially) {
       dispatch('track')
+    }
+    dispatch('setupTooltip')
+  },
+  setupTooltip({ getters, rootGetters }) {
+    if (getters.showTooltip) {
+      const { map } = rootGetters
+      const overlay = new Overlay({
+        element: getTooltip(),
+        positioning: 'bottom-center',
+        offset: [0, -5],
+      })
+      map.addOverlay(overlay)
+      map.on('pointermove', ({ pixel, dragging }) => {
+        if (dragging) {
+          return
+        }
+        const features = map.getFeaturesAtPixel(pixel, {
+          layerFilter: (layer) =>
+            layer.get('name') === 'geoLocationMarkerLayer',
+        })
+
+        const coordinate = features.length
+          ? map.getCoordinateFromPixel(pixel)
+          : undefined
+        overlay.setPosition(coordinate)
+      })
     }
   },
 
@@ -173,7 +201,7 @@ const actions: PolarActionTree<GeoLocationState, GeoLocationGetters> = {
         image: new Icon({
           // TODO: It might be interesting to be able to change the color.
           src: `data:image/svg+xml;utf8,${geoLocationMarker}`,
-          scale: 0.09,
+          scale: 0.08,
           opacity: 1,
         }),
       })