+++ packages/types/custom/core.ts
@@ -254,6 +254,7 @@ export interface GeoLocationConfiguration extends LayerBoundPluginOptions {
   checkLocationInitially: boolean
   /** whether to keep center on user or allow movement after first zoom to */
   keepCentered: boolean
+  showTooltip?: boolean
   /**
    * Limits the viewable GFIs per layer by this number. The first n elements
    * are chosen arbitrarily. Useful if you e.g. just want one result, or to