+++ packages/plugins/GeoLocation/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Feature: Update`@polar/lib-passes-boundary-check` to major version 2. (TODO: Publish and install library before publishing this update in a new version.)
 - Feature: Add `boundaryOnError` parameter to let user define behaviour on boundary check errors.
+- Feature: Add optional tooltip to user geolocation map icon.
 
 ## 1.0.0