+++ packages/types/custom/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Feature: Add parameter `boundaryOnError` to `PinsConfiguration`.
 - Feature: Add parameter `boundaryOnError` to `GeoLocationConfiguration`.
+- Feature: Add flag `showTooltip` to `GeoLocationConfiguration`.
 - Feature: Add type for CoreState's `center` field.
 - Feature: Remodel type structure to deduplicate fields now modeled in `LayerBoundPluginOptions`.
 - Feature: Add new optional parameter `initial` to `PinsConfiguration` including related interface `InitialPin`.